refactor(evm): rename proxy_deploy_contract module to deploy_contract for clarity
